+++ composer.json
@@ -13,6 +13,7 @@
         "fideloper/proxy": "^4.0",
         "intervention/image": "^2.5",
         "laravel/framework": "^8.0.0",
+        "laravel/legacy-factories": "^1.3",
         "laravel/scout": "^9.8",
         "laravel/telescope": "^4.10",
         "laravel/tinker": "^2.0",
